+++ packages/pluggable-widgets-tools/configs/rollup.config.mjs
@@ -13,7 +13,7 @@ import typescript from "@rollup/plugin-typescript";
 import colors from "ansi-colors";
 import postcssImport from "postcss-import";
 import postcssUrl from "postcss-url";
-import { loadConfigFile } from "rollup/dist/loadConfigFile.js";
+import loadConfigFile from "rollup/dist/loadConfigFile.js";
 import clear from "rollup-plugin-clear";
 import command from "rollup-plugin-command";
 import license from "rollup-plugin-license";